# ddmedistock/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDMediStock API",
    "version": "1.0.0",
    "description": "B2B medical procurement & RFQ platform API. Authentication is via an httpOnly session cookie (`ddms_session`) issued by POST /api/auth/login. State-changing requests are CSRF-protected (same-origin) and rate-limited."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in and receive a session cookie",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "role": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a customer organization + user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "name",
                  "organizationName"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "name": {
                    "type": "string"
                  },
                  "organizationName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Search the product catalog",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Free-text query"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rfqs": {
      "post": {
        "tags": [
          "RFQ"
        ],
        "summary": "Create an RFQ (customer) — triggers AI parse + match",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "rawText"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "rawText": {
                    "type": "string",
                    "description": "Raw RFQ text / pasted requirement sheet"
                  },
                  "sourceType": {
                    "type": "string",
                    "enum": [
                      "TEXT",
                      "CSV",
                      "EXCEL",
                      "PDF",
                      "IMAGE"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RFQ created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQ"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/quotations/{id}": {
      "patch": {
        "tags": [
          "Quotations"
        ],
        "summary": "Edit a quotation's items/notes (staff, not when published)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Published quotations cannot be edited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/quotations/{id}/approval": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "summary": "Workflow & approval history",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "History"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Quotations"
        ],
        "summary": "Submit for approval / approve / reject (staff)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "SUBMIT",
                      "APPROVE",
                      "REJECT"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied"
          },
          "403": {
            "description": "Capability denied (e.g. AI/customer cannot approve)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Illegal workflow transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ddms_session"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "uom": {
            "type": "string"
          },
          "originType": {
            "type": "string",
            "enum": [
              "DOMESTIC",
              "IMPORTED"
            ]
          }
        }
      },
      "RFQ": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "QuotationEdit": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "customerNotes": {
            "type": "string"
          },
          "internalNotes": {
            "type": "string"
          },
          "taxPct": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_APPROVAL"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "description"
              ],
              "properties": {
                "description": {
                  "type": "string"
                },
                "quantity": {
                  "type": "number"
                },
                "unit": {
                  "type": "string"
                },
                "unitCost": {
                  "type": "number"
                },
                "marginPct": {
                  "type": "number"
                },
                "unitPrice": {
                  "type": "number"
                },
                "taxPct": {
                  "type": "number"
                },
                "originType": {
                  "type": "string",
                  "enum": [
                    "DOMESTIC",
                    "IMPORTED"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}